submissions/2026-04-29__001_synthetic_mine_throughput__claude-code__claude-opus-4-7__plan-mode/requirements.txt
simpy>=4.0
numpy>=1.24
pandas>=2.0
scipy>=1.10
networkx>=3.0
pyyaml>=6.0
matplotlib>=3.7
